essai_dccstore/app/requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
dash==2.7.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.10.2
dash-table==5.0.0
dash-tools==1.10.7
distlib==0.3.6
docopt==0.6.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.2.0
gunicorn==20.1.0
identify==2.5.12
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
miniful==0.0.6
mpld3==0.5.9
nltk==3.7
nodeenv==1.7.0
numpy==1.21.6
packaging==21.0
pandas==1.3.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.10.0
pluggy==1.0.0
pre-commit==2.16.0
psycopg2-binary==2.9.5
py==1.11.0
pyFUME==0.2.25
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scipy==1.7.3
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
tenacity==8.1.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.13
virtualenv==20.17.1
visdcc==0.0.50
Werkzeug==2.2.2
wordcloud==1.8.2.2
yarg==0.1.9
yolk3k==0.9
